Setup of the lexer's allocator
- Replaced `long int` by `size_t` for `i` and `j` indexes.
- Initialized each structure.
- Fixed the issue where some pointers to tokens were not updated after a
  reallocation.

+++ headers/parser_allocation.h
@@ -1,10 +1,11 @@
 #ifndef PARSER_ALLOCATION
 #define PARSER_ALLOCATION
 
+#include <stdlib.h>
 #include "parser_def.h"
 
 bool parser_allocate_chunk(
-	long int minimum,
+	size_t minimum,
 	Parser* parser);
 
 #endif
