+++ headers/parser_identifier.h
@@ -5,8 +5,8 @@
 #include "parser_def.h"
 
 int if_identifier_create_nodes(
-	long int* i,
-	long int* j,
+	size_t* i,
+	size_t* j,
 	MemoryArea* memArea,
 	Parser* parser);
