+++ src/parser/parser_allocation.c
@@ -3,11 +3,11 @@
 #include "parser_allocation.h"
 
 bool parser_allocate_chunk(
-long int minimum,
+size_t minimum,
 Parser* parser) {
 #define NODES_CHUNK 4096
 	if(parser->count <= minimum) {
-		const long int reserve = minimum / NODES_CHUNK + 1;
+		const size_t reserve = minimum / NODES_CHUNK + 1;
 		Node* nodes_realloc = realloc(
 			parser->nodes,
 			reserve * NODES_CHUNK * sizeof(Node));
